fix: resolve 8 bugs across codebase
- SatelliteCacheStore: use独立 SharedPreferences 文件名，避免与 SettingsStore key 冲突
- SettingsStore: GPS 坐标改用 String 存储 Double，修复 Float 精度丢失
- WeatherCard: 温度格式化指定 Locale.US，避免不同 locale 下小数点变逗号
- WeatherApiService: 移除 3 处冗余嵌套 withContext(Dispatchers.IO)
- SatellitePredictor: catch Exception 前显式捕获 CancellationException 向上传播
- MainViewModel: dailyQuoteDayOfYear 持久化到 SettingsStore，避免重启后重复请求
- build.gradle.kts: encryptSecrets JSON 拼接添加 escapeJson() 转义
- MainActivity: BackgroundLayer 中 remember(uri) 改为 remember，LaunchedEffect 先 recycle 旧 Bitmap 再加载新的，修复 native 内存泄漏

+++ app/src/main/java/com/example/radioarealocator/data/reminder/ReminderRefreshWorker.kt
@@ -57,8 +57,8 @@ class ReminderRefreshWorker(
                 return Result.success()
             }
 
-            val latitude = settingsStore.lastLatitude.toDouble()
-            val longitude = settingsStore.lastLongitude.toDouble()
+            val latitude = settingsStore.lastLatitude
+            val longitude = settingsStore.lastLongitude
             val satelliteSource = settingsStore.satelliteSource
 
             // 获取 TLE 数据：优先使用缓存，超过 24 小时则重新下载
